+++ math/src/main/java/smile/math/blas/Side.java
@@ -23,23 +23,23 @@ package smile.math.blas;
  */
 public enum Side {
     /** A * B */
-    LEFT ((byte) 141, (byte) 'L'),
+    LEFT (141, (byte) 'L'),
     /** B * A */
-    RIGHT((byte) 142, (byte) 'R');
+    RIGHT(142, (byte) 'R');
 
     /** Byte value passed to BLAS. */
-    private final byte blas;
+    private final int blas;
     /** Byte value passed to LAPACK. */
     private final byte lapack;
 
     /** Constructor. */
-    Side(byte blas, byte lapack) {
+    Side(int blas, byte lapack) {
         this.blas = blas;
         this.lapack = lapack;
     }
 
     /** Returns the byte value for BLAS. */
-    public byte blas() { return blas; }
+    public int blas() { return blas; }
 
     /** Returns the byte value for LAPACK. */
     public byte lapack() { return lapack; }
